refactor(player): E.32 — extract UI state actions as factory
Ten pure-UI state setters move into `src/store/uiStateActions.ts` as a
`createUiStateActions(set)` factory:

  - `setStarredOverride`, `setUserRatingOverride` — optimistic overrides
  - `openContextMenu`, `closeContextMenu` — context menu modal
  - `openSongInfo`, `closeSongInfo` — song info modal
  - `toggleQueue`, `setQueueVisible` — queue panel toggle with persisted
    localStorage round-trip
  - `toggleFullscreen` — fullscreen player toggle
  - `toggleRepeat` — repeat mode cycle

All ten are pure state mutators (no audio engine / network calls).
Store body uses `...createUiStateActions(set)`. Action-factory pattern
established by E.31 reused here. `persistQueueVisibility` import drops
from playerStore.

playerStore 1504 → 1463 LOC.
